requirements.txt
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0  # Para cargar variables de entorno desde .env
MetaTrader5>=5.0.45  # Para conexión con MetaTrader 5
requests>=2.31.0  # Para alertas de Telegram
# python-binance>=1.0.19  # Descomenta si usas Binance
# ib-insync>=0.9.86  # Descomenta si usas Interactive Brokers

